+++ src/Demo/Logitar.Demo.Ui/Domain/Events/TodoDeletedEvent.cs
@@ -1,0 +1,11 @@
+﻿using Logitar.EventSourcing;
+
+namespace Logitar.Demo.Ui.Domain.Events;
+
+public record TodoDeletedEvent : DomainEvent
+{
+  public TodoDeletedEvent()
+  {
+    IsDeleted = true;
+  }
+}