+++ src/Demo/Logitar.Demo.Ui/Domain/Events/TodoUpdatedEvent.cs
@@ -1,0 +1,9 @@
+﻿using Logitar.EventSourcing;
+
+namespace Logitar.Demo.Ui.Domain.Events;
+
+public record TodoUpdatedEvent : DomainEvent
+{
+  public string? Text { get; set; }
+  public bool? IsDone { get; set; }
+}